+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+### Added
+- `core` - Add `getRecoveredData` method to return helper data
+- `react` - Add `helperData` parameter to `onSuccess` callback
 
 ## [0.2.0] - 2025-08-05
 ### Added